refactor(modal): split use-modal-stack into focused composables
Replaces the single use-modal-stack.ts monolith with four colocated
files: use-modal-guards.ts (scroll lock + esc-to-close), use-modal-
recede.ts (recede/restore choreography), use-modal-transitions.ts
(enter/leave animations), and mobile-below.ts (shared mobile-breakpoint
defaults/isMobileFor helper).

Also fixes an audit finding: recede/restore now derives "receded"
structurally as every entry except the current top, recomputed off the
full stack each time it changes, instead of assuming stack length
changes by exactly one push/pop at a time — a batch of opens/closes
now settles correctly.
