+++ src/components/ui-kit/modal/mode-config.ts
@@ -24,8 +24,8 @@ export const MODAL_MODE_CONFIG: Record<ModalMode, ModeConfig> = {
   // The `mobile-modal:` Tailwind variant (defined in
   // `src/styles/custom-variants.css`) activates these utilities when the
   // viewport drops below the modal's threshold (data-mobile-below-width/height
-  // are stamped by modal.vue). Keeping the class string static on the Vue side
-  // avoids touch-disrupting setAttribute calls during scroll on iOS Safari.
+  // are stamped by ui-kit/modal/index.vue). Keeping the class string static on
+  // the Vue side avoids touch-disrupting setAttribute calls during scroll on iOS Safari.
   'mobile-sheet': {
     containerClass:
       'items-center mobile-modal:flex-col mobile-modal:overflow-y-auto mobile-modal:overscroll-y-contain mobile-modal:justify-start mobile-modal:pt-4 mobile-modal:pointer-events-auto',
